💅🏻 Refactoring: Votes
