--- src/components/Comment/Votes/index.js
@@ -1,3 +1,0 @@
-import Votes from "./Votes";
-
-export default Votes;
